% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{AddSickleJrMetadata}
\alias{AddSickleJrMetadata}
\title{Add metadata to an object of class SickleJr}
\usage{
AddSickleJrMetadata(SickleJr, metadata, metadataname)
}
\arguments{
\item{SickleJr}{An object of class SickleJr holding at least one count matrix of omics data}

\item{metadata}{Metadata to add to the SickleJr object; there are no restrictions on type}

\item{metadataname}{A string input that indicates the desired name for the added
metadata.}
}
\value{
An object of class SickleJr with added metadata
}
\description{
Add any type of metadata to an object of class SickleJr. Metadata
are stored in list format under the name specified in \code{metadataname} of each node in slot \code{metadata}.
}
\examples{
SimSickleJrSmall<-AddSickleJrMetadata(SimSickleJrSmall,
SimData$cell_type,"cell_types_full_data")
}
