% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{CreateSickleJr}
\alias{CreateSickleJr}
\title{Create an object of type SickleJr}
\usage{
CreateSickleJr(count.matrices, names = NULL)
}
\arguments{
\item{count.matrices}{A list of count matrices; each view is one entry in the
list. These count matrices should already be QC'd and the features filtered}

\item{names}{Optional parameter with names for the count matrices}
}
\value{
A SickleJr object with sparse count matrices
}
\description{
Creates an object of type SickleJr that includes lists of sparse
count matrices and allows users to specify the names of those count matrices
}
\examples{
SickleJrSim<-CreateSickleJr(SimData$Xmatrices)
}
