% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lossCalcRWrapper}
\alias{lossCalcRWrapper}
\title{Calculate the loss function in R}
\usage{
lossCalcRWrapper(datamatL, WL, H, AdjL, DL, lambdaWL, lambdaH, diffFunc)
}
\arguments{
\item{datamatL}{An R list where each entry contains a normalized, sparse
\eqn{\mathbf{X}^v} matrix corresponding to single-cell modality \eqn{v}}

\item{WL}{An R list containing initialized values of the \eqn{\mathbf{W}^v}
within each modality \eqn{v}}

\item{H}{A matrix containing initialized values for the shared
\eqn{\mathbf{H}}}

\item{AdjL}{An R list containing all of the adjacency matrices for the
feature-feature similarity graphs in sparse format;  note that
\eqn{\mathtt{D-Adj}} is the graph Laplacian}

\item{DL}{An R list containing all of the degree matrices of the
feature-feature similarity graphs;  note that \eqn{\mathtt{D-Adj}} is the
graph Laplacian}

\item{lambdaWL}{A list of the \eqn{\lambda_{\mathbf{W}^v}} corresponding to
modality \eqn{v}}

\item{lambdaH}{A double containing the desired value for
\eqn{\lambda_{\mathbf{H}}}}

\item{diffFunc}{A vector of strings indicating what type of divergence to
use;  set to the Poisson Kullback-Leibler divergence
(\eqn{\mathtt{``klp"}}) by default, but the Frobenius norm
(\eqn{\mathtt{``fr"}}) is also available}
}
\value{
A double containing the value of the jrSiCKLSNMF loss function
}
\description{
This is a wrapper function that allows for the calculation of
the loss function in R code.
}
