% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{MinibatchDiagnosticPlot}
\alias{MinibatchDiagnosticPlot}
\title{Plot a diagnostic plot for the mini-batch algorithm}
\usage{
MinibatchDiagnosticPlot(SickleJr)
}
\arguments{
\item{SickleJr}{An object of class SickleJr}
}
\value{
An object of class SickleJr with mini-batch diagnostic plots added to the \code{plots} slot.
}
\description{
To ensure sufficient convergence of the loss for jrSiCKLSNMF with mini-batch updates, we
plot the loss vs the number of iterations for the mini-batch algorithm.
After a certain number of iterations, the loss should appear to oscillate
around a value. Before continuing with downstream analyses, please ensure that
the loss exhibits this sort of behavior. For the mini-batch algorithm, it is not
possible to use the convergence criteria used for the batch version of the
algorithm.
}
\examples{
SimSickleJrSmall<-MinibatchDiagnosticPlot(SimSickleJrSmall)
}
