% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{OnlineDiagnosticPlot}
\alias{OnlineDiagnosticPlot}
\title{Plot a diagnostic plot for the online algorithm}
\usage{
OnlineDiagnosticPlot(SickleJr)
}
\arguments{
\item{SickleJr}{A SickleJr Object}
}
\value{
A diagnostic plot. SickleJr object remains unchanged.
}
\description{
Plot the loss vs the number of iterations for the online algorithm. After a certain number of iterations,
the loss should appear to oscillate around a value.
}
