% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{setLambdas}
\alias{setLambdas}
\title{setLambdas}
\usage{
setLambdas(SickleJr, lambdaWlist, lambdaH = 0)
}
\arguments{
\item{SickleJr}{a SickleJr object}

\item{lambdaWlist}{A list of graph regularization constraints for the W matrices.}

\item{lambdaH}{lambda H value to add. Defaults to 0.}
}
\value{
SickleJr object with added lambda values
}
\description{
Set the lambda values for a SickleJr object
}
