% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{AddSickleJrMetadata}
\alias{AddSickleJrMetadata}
\alias{AddSickleJrMetaData}
\title{AddSickleJrMetadata}
\usage{
AddSickleJrMetaData(SickleJr, metadata, metadataname)
}
\arguments{
\item{SickleJr}{a SickleJr object}

\item{metadata}{The metadata you wish to add to the object}

\item{metadataname}{The name that you wish to call the added metadata}
}
\value{
SickleJr with added metadata
}
\description{
Add metadata to a SickleJr object
}
