% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{BuildKNNGraphLaplacians}
\alias{BuildKNNGraphLaplacians}
\title{BuildKNNGraphLaplacians}
\usage{
BuildKNNGraphLaplacians(SickleJr, k = 20)
}
\arguments{
\item{SickleJr}{A SickleJr object}

\item{k}{Number of KNN neighbors to calculate. By default, is set to 20}
}
\value{
A list of graph Laplacians
}
\description{
Generate graph Laplacians for graph regularization of SickleJrNMF from the list of count matrices
}
