% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{BuildSNNGraphLaplacians}
\alias{BuildSNNGraphLaplacians}
\title{Build SNN graphs and generate their graph Laplacians}
\usage{
BuildSNNGraphLaplacians(SickleJr, k = 20)
}
\arguments{
\item{SickleJr}{A SickleJr object}

\item{k}{Number of KNN neighbors to calculate. By default, is set to 20}
}
\value{
A list of graph Laplacians in sparse matrix format
}
\description{
Generate graph Laplacians for graph regularization of
jrSiCKLSNMFNMF from the list of normalized count matrices
}
