% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{CalculateUMAPSickleJr}
\alias{CalculateUMAPSickleJr}
\title{CalculateUMAPSickleJr}
\usage{
CalculateUMAPSickleJr(
  SickleJr,
  umap.settings = umap::umap.defaults,
  view = NULL
)
}
\arguments{
\item{SickleJr}{An object of class SickleJr}

\item{umap.settings}{Optional settings for UMAP. Set to default.}

\item{view}{If this parameter is set, this will perform UMAP on W_view%*%t(H) rather than
on H alone.}
}
\value{
a SickleJr with added UMAP calculations based on the H matrix alone or within a view
}
\description{
Perform UMAP on the H matrix alone (default) or within a view
}
