% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{ClusterSickleJr}
\alias{ClusterSickleJr}
\title{ClusterSickleJr}
\usage{
ClusterSickleJr(SickleJr, numclusts, method = "kmeans", neighbors = NULL)
}
\arguments{
\item{SickleJr}{an object of class SickleJr}

\item{numclusts}{Number of clusters}

\item{method}{Clustering method. Can choose between kmeans and spectral clustering}

\item{neighbors}{parameter for spectral clustering}
}
\value{
a SickleJr object with added clustering information
}
\description{
Perform either K-means or spectral clustering on the H matrix.
Defaults to kmeans.
}
