% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{createSickleJr}
\alias{createSickleJr}
\title{createSickleJr}
\usage{
createSickleJr(count.matrices)
}
\arguments{
\item{count.matrices}{A count matrix that has already been filtered such that all genes and peaks
appear in at least 10 cells.}
}
\value{
A SickleJr object
}
\description{
creates an object of type SickleJr
}
