% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{determineClusters}
\alias{determineClusters}
\title{Perform clustering diagnostics}
\usage{
determineClusters(
  SickleJr,
  numclusts = 2:20,
  clusteringmethod = "kmeans",
  diagnosticmethods = c("wss", "silhouette", "gap_stat"),
  clValidvalidation = "internal",
  printDiagnosticplots = TRUE,
  printClValidDiagnostics = TRUE
)
}
\arguments{
\item{SickleJr}{A SickleJr object}

\item{numclusts}{a vector of clusters to test}

\item{clusteringmethod}{Clustering method. Defaults to k-means. Since other methods
are not implemented in jrSiCKLSNMF, it is recommended to use k-means.}

\item{diagnosticmethods}{Which methods to plot. Defaults to all three of the available: wss, silhouette, and gap_stat}

\item{clValidvalidation}{validation method to use for ClValid. Defaults to internal.}

\item{printDiagnosticplots}{Whether to print the diagnostic plots}

\item{printClValidDiagnostics}{Whether to print the diagnostic results from clValid}
}
\value{
a SickleJr object with cluster diagnostics
}
\description{
A wrapper for clValid and factoextra functions to perform clustering diagnostics
}
