% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{GenerateWmatricesandHmatrix}
\alias{GenerateWmatricesandHmatrix}
\title{GenerateWmatricesandHmatrix}
\usage{
GenerateWmatricesandHmatrix(SickleJr, d = 10, rowReg = "L2Norm")
}
\arguments{
\item{SickleJr}{a SickleJr object}

\item{d}{number of latent factors to use. Defaults to 10.}

\item{rowReg}{Choose whether or not to enforce constraints on the rows. Enter "None" for no constraints
Defaults to L2Norm.}
}
\value{
SickleJr A SickleJr object with W and H matrices added.
}
\description{
Perform normalization for count data. If you are planning to use the Frobenius norm,
set frob=TRUE
}
