% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jrSiCKLSNMF}
\alias{jrSiCKLSNMF}
\title{jrSiCKLSNMF}
\usage{
jrSiCKLSNMF(
  datamatL,
  WL,
  H,
  AdjL,
  DL,
  lambdaWL,
  lambdaH,
  initsamp,
  suppress_warnings,
  diffFunc = "klp",
  Hconstraint = "None",
  differr = 1e-06,
  rounds = 1000,
  display_progress = TRUE,
  online = TRUE,
  batchsize = 100L,
  random_W_updates = TRUE,
  minrounds = 100L
)
}
\arguments{
\item{datamatL}{An R list where each entry contains a sparse X matrix corresponding to a single cell data view
Each X is m^v features by n cells. Features can differ across matrices; however, n must be the same for
each view. All data are measured on the same set of cells}

\item{WL}{An R list containing initialized values for the W within each view. These are passed by reference}

\item{H}{A matrix containing initialized values for the shared H}

\item{AdjL}{An R list containing all of the adjacency matrices for the
feature-feature similarity graphs in sparse format. Note that D-Adj is the
graph Laplacian}

\item{DL}{An R list containing all of the degree matrices of the
feature-feature similarity graphs. Note that D-Adj is the graph Laplacian}

\item{lambdaWL}{A list of each lambdaW for each view}

\item{lambdaH}{A double containing the desired value for H}

\item{initsamp}{A vector of randomly selected rows of H on which to run the objective function}

\item{suppress_warnings}{Indicates whether warnings should be suppressed}

\item{diffFunc}{A string indicating what type of divergence to use. It is Poisson Kulback Leibler by default}

\item{Hconstraint}{A string that indicates whether you want to force constraints on the rows of H. Enter "None" for
no constraints, enter "L1Norm" to ensure all rows of H sum to 1, and "L2Norm" to ensure that the the
L2 norm of each row of H equals 1. Please note that jrSiCKLSNMF stores H as the transpose of H, so the code
will perform regularization on the columns of the transpose of H.}

\item{differr}{A double containing the tolerance}

\item{rounds}{A double containing the number of rounds}

\item{display_progress}{A boolean to indicate whether the user wants to display the progress}

\item{online}{A boolean to indicate whether the user would like to use the online version of the algorithm.
The online version}

\item{batchsize}{Number of batches for online updates}

\item{random_W_updates}{Indicates whether or not to use a random_W_updates algorithm for updating
the W lists. If true, will only update W every 5 iterations.}

\item{minrounds}{A minimum number of rounds for the algorithm to run. Most useful for the online algorithm}
}
\value{
An R list containing values for the objective function.
}
\description{
Perform joint non-negative matrix factorization (NMF) across multiple views of single cell data.
Users can choose to use the Poisson Kullback-Leibler divergence or the Frobenius norm.
Users can also set graph regularization constraints on W and sparsity constraints on H.
This function updates WL and H.
}
