% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{NormalizeCountMatrices}
\alias{NormalizeCountMatrices}
\title{Normalize the count matrices and set whether to use the KL divergence
based on the Poisson distribution or the Frobenius norm}
\usage{
NormalizeCountMatrices(SickleJr, diffFunc = "klp", scaleFactor = NULL)
}
\arguments{
\item{SickleJr}{An object of type SickleJr with a count matrix. Users can choose to normalize using median
library size normalization or log(x+1) normalization}

\item{diffFunc}{A string. Set to "klp" to use the KL divergence based on the Poisson distribution
or to "fr" to use the Frobenius norm. Defaults to KL divergence. This also determines
the type of normalization to use.}

\item{scaleFactor}{A list of numeric values to use as scale factors for the log(x+1)
normalization when using the Frobenius diffFunction}
}
\value{
A list of sparse, normalized matrices
}
\description{
Perform normalization for count data. If you are planning to use the Frobenius norm,
set frob=TRUE to log(x+1) normalize your count data. This step can be skipped for percentage data and
spectra data. You may also skip this if you would like to perform a different form
of normalization; however, please ensure that if using the Frobenius norm that all of the values are
divided by the sum of all of the values.
}
