% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{PlotLossvsLatentFactors}
\alias{PlotLossvsLatentFactors}
\title{Creating plots to help determine the number of latent factors to use for jrSiCKLSNMF}
\usage{
PlotLossvsLatentFactors(
  SickleJr,
  rounds = 200,
  differr = 1e-05,
  d_vector = c(2:20),
  parallel = FALSE,
  nCores = detectCores() - 1,
  subsampsize = NULL,
  online = FALSE,
  random = FALSE,
  random_W_updates = FALSE,
  seed = NULL,
  batchsize = -1,
  lossonsubset = FALSE,
  losssubsetsize = dim(SickleJr@count.matrices[[1]])[2]
)
}
\arguments{
\item{SickleJr}{A SickleJr object}

\item{rounds}{Number of rounds to use. Defaults to 20. This process will take some time,
so a high number of rounds is not recommended}

\item{differr}{A tolerance of updates after which to stop updating. For these plots,
this is set to 1e-4 by default}

\item{d_vector}{Vector of D values to test over. Defaults from 2 to 20.}

\item{parallel}{Indicates whether or not you want to perform this using
parallel computing}

\item{nCores}{Number of desired cores. If null, defaults to the number of cores
minus 1 for user convenience.}

\item{subsampsize}{Indicates whether you want to perform this on a random subsample rather
than on the whole dataset. Will speed up the process but will have lower accuracy on desired
number of d}

\item{online}{Indicates whether you want an online algorithm. Defaults to FALSE}

\item{random}{Indicates whether to use random initialization to generate the W and H matrices
Defaults to FALSE as this in general is less accurate.}

\item{random_W_updates}{Only used for online algorithms. If TRUE, only updates W using the second
to last H subset on the online algorithm}

\item{seed}{Set a random seed}

\item{batchsize}{Desired batch size. Do not use if you are using a subsample.}

\item{lossonsubset}{Indicates whether to calculate the loss on a subset rather than the full dataset.
Speeds up computation for larger datasets.}

\item{losssubsetsize}{Number of cells to use for the loss subset. Defaults to full number of cells}
}
\value{
A SickleJr with an added ggplot object in slot LatentFactorDeterminationPlot
}
\description{
This generates plots of the lowest achieved loss after a
user-specified number of iterations (default 200)
of the jrSiCKLSNMF algorithm for each latent factor (defaults to 2:20). This operates similarly to a
scree plot, so please select a number of latent factors that corresponds to the
elbow of the plot. This function will not uniformly decrease, so an increase in loss
may be expected. Select the minimum loss or a point that appears to be an elbow.
}
