% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{plotSickleJrUMAP}
\alias{plotSickleJrUMAP}
\title{plotSickleJrUMAP}
\usage{
plotSickleJrUMAP(
  SickleJr,
  umap.view = "H",
  cluster = "kmeans",
  title = "",
  colorbymetadata = NULL,
  legendname = NULL
)
}
\arguments{
\item{SickleJr}{An object of type SickleJr}

\item{umap.view}{String corresponding to the name of the UMAP you are interested in. Defaults to H}

\item{cluster}{Cluster you wish to color by}

\item{title}{Optional title}

\item{colorbymetadata}{Name of metadata column if you wish to color by metadata}

\item{legendname}{Option if you want to give a different name for your legend}
}
\value{
A ggplot2 object of the plots of UMAP1 and UMAP2 for the SickleJr object
}
\description{
Plot first and second dimensions of UMAP and color either by cluster or metadata
}
