% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{RunjrSiCKLSNMF}
\alias{RunjrSiCKLSNMF}
\title{RunjrSiCKLSNMF}
\usage{
RunjrSiCKLSNMF(
  SickleJr,
  rounds = 300,
  differr = 1e-06,
  diffFunc = "klp",
  display_progress = TRUE
)
}
\arguments{
\item{SickleJr}{a SickleJr object}

\item{rounds}{Number of rounds. Defaults to 300.}

\item{differr}{Tolerance for updating. Defaults to 1e-6.}

\item{diffFunc}{choose between "klp" for Poisson-based kullback leibler divergence}

\item{display_progress}{whether to display the progress bar for jrSiCKLSNMF}
}
\value{
a SickleJr object with updated W matrices, updated H matrices, and a vector of values for
the loss function
}
\description{
Wrapper function to run jrSiCKLSNMF on a SickleJr object. Performs jrSiCKLSNMF on
the given SickleJr
}
