% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{SetLambdasandRowReg}
\alias{SetLambdasandRowReg}
\title{Set lambda values, type of row regularization, for a SickleJr object}
\usage{
SetLambdasandRowReg(
  SickleJr,
  lambdaWlist = list(10, 50),
  lambdaH = 500,
  rowReg = "None"
)
}
\arguments{
\item{SickleJr}{a SickleJr object}

\item{lambdaWlist}{A list of graph regularization constraints for the W matrices.
Defaults to 2 views with the RNA view equal to 10 and the ATAC view equal to 50}

\item{lambdaH}{lambda H value to add. Defaults to 500.}

\item{rowReg}{Choose whether or not to enforce constraints on the rows. Enter
"None" for no constraints and "L2Norm" for an L2 Norm constraint on the rows
of H. Defaults to "None"}
}
\value{
SickleJr object with added lambda values
}
\description{
Provide the values for the graph regularization \eqn{\lambda_{\textbf{W}^v}}
for each view as a list and provide a
}
