% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrSiCKLSNMF.R
\name{SetWandHfromWHinitials}
\alias{SetWandHfromWHinitials}
\title{Set W matrices and H matrix from pre-calculated values}
\usage{
SetWandHfromWHinitials(SickleJr, d)
}
\arguments{
\item{SickleJr}{A SickleJr object}

\item{d}{The number of desired latent factors}
}
\value{
a SickleJr object with a filled in W and H matrices.
}
\description{
Use values calculated in the step to determine number of latent factors in the initial
steps for the jrSiCKLSNMF algorithm. If only a subset was calculated, this uses
nndsvd to calculate missing H values
}
